Change the structure to separate tests from application
Wk4
