+++ src/main/java/iterator/Subject.java
@@ -1,4 +1,4 @@
-package com.iterator.lessons;
+package main.java.iterator;
 
 import java.util.Iterator;
